+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Updates `describe_class` to account for RSpecs `:system` wrapper of rails system tests. ([@EliseFitz15][])
 * Add `RSpec/ExpectChange` cop to enforce consistent usage of the change matcher. ([@Darhazer][])
+* Add autocorrect support to `RSpec/LetBeforeExamples`. ([@Darhazer][])
 
 ## 1.21.0 (2017-12-13)